+++ src/youtube_search_engine.py
@@ -1,0 +1,89 @@
+
+
+import urllib.request
+import pafy
+from googleapiclient.discovery import build
+from googleapiclient.errors import HttpError
+from oauth2client.tools import argparser
+
+
+#API Key for YouTube and KS Search Engine
+google_cloud_api_key='ENTER-YOUR-GOOGLE-CLOUD-API-KEY-HERE'
+
+#YouTube API Constants
+DEVELOPER_KEY = google_cloud_api_key
+YOUTUBE_API_SERVICE_NAME = 'youtube'
+YOUTUBE_API_VERSION = 'v3'
+
+
+
+
+#Function to search YouTube and get videoid
+def youtube_search(query):
+  youtube = build(YOUTUBE_API_SERVICE_NAME, YOUTUBE_API_VERSION,
+    developerKey=DEVELOPER_KEY)
+
+  req=query
+  # Call the search.list method to retrieve results matching the specified
+  # query term.
+  search_response = youtube.search().list(
+    q=query,
+    part='id,snippet'
+  ).execute()
+  #print(search_response)
+
+  videos = []
+  channels = []
+  playlists = []
+  videoids = []
+  channelids = []
+  playlistids = []
+
+  # Add each result to the appropriate list, and then display the lists of
+  # matching videos, channels, and playlists.
+
+  for search_result in search_response.get('items', []):
+
+    if search_result['id']['kind'] == 'youtube#video':
+      videos.append('%s (%s)' % (search_result['snippet']['title'],
+                                 search_result['id']['videoId']))
+      videoids.append(search_result['id']['videoId'])
+
+    elif search_result['id']['kind'] == 'youtube#channel':
+      channels.append('%s (%s)' % (search_result['snippet']['title'],
+                                   search_result['id']['channelId']))
+      channelids.append(search_result['id']['channelId'])
+
+    elif search_result['id']['kind'] == 'youtube#playlist':
+      playlists.append('%s (%s)' % (search_result['snippet']['title'],
+                                    search_result['id']['playlistId']))
+      playlistids.append(search_result['id']['playlistId'])
+
+  #Results of YouTube search. If you wish to see the results, uncomment them
+  # print 'Videos:\n', '\n'.join(videos), '\n'
+  # print 'Channels:\n', '\n'.join(channels), '\n'
+  # print 'Playlists:\n', '\n'.join(playlists), '\n'
+
+  #Checks if your query is for a channel, playlist or a video and changes the URL accordingly
+  if 'channel'.lower() in  str(req).lower() and len(channels)!=0:
+      urlid=channelids[0]
+      YouTubeURL=("https://www.youtube.com/watch?v="+channelids[0])
+  elif 'playlist'.lower() in  str(req).lower() and len(playlists)!=0:
+      urlid=playlistids[0]
+      YouTubeURL=("https://www.youtube.com/watch?v="+playlistids[0])
+  else:
+      urlid=videoids[0]
+      YouTubeURL=("https://www.youtube.com/watch?v="+videoids[0])
+
+  return YouTubeURL,urlid
+
+
+#Function to get streaming links for YouTube URLs
+def youtube_stream_link(videourl):
+    url=videourl
+    video = pafy.new(url)
+    bestvideo = video.getbest()
+    bestaudio = video.getbestaudio()
+    audiostreaminglink=bestaudio.url
+    videostreaminglink=bestvideo.url
+    return audiostreaminglink,videostreaminglink
